Replace Default handler with action-mixins (for better reusability)
